requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
torch
torchvision
argparse
Pillow
requests
tqdm
open_clip_torch
git+https://github.com/huggingface/transformers.git
git+https://github.com/pipinstallyp/BLIP
salesforce-lavis
git+https://git@github.com/seatgeek/thefuzz.git@0.19.0#egg=thefuzz
python-Levenshtein 
accelerate
tiktoken
einops
transformers_stream_generator==0.0.4
scipy
tensorboard
matplotlib